Refs #109 Refactored factory of SiteBaker instances
Public API is now much cleaner.
