+++ conf/yawg-defaults.conf
@@ -19,7 +19,7 @@
 # MUST run devtools/bin/bild-sync-version to propagate the change to
 # other source files.
 #
-YAWG_MAIN_VERSION=0.0.1
+YAWG_MAIN_VERSION=0.1.0
 
 
 #
