+++ modules/core/src/main/java/com/varmateo/yawg/freemarker/FreemarkerDataModel.java
@@ -6,8 +6,6 @@
 
 package com.varmateo.yawg.freemarker;
 
-import java.util.Optional;
-
 import com.varmateo.yawg.PageTemplateDataModel;
