+++ modules/pom.xml
@@ -11,7 +11,7 @@ Copyright (c) 2015-2016 Jorge Nunes, All Rights Reserved.
 
   <groupId>com.varmateo.yawg</groupId>
   <artifactId>yawg-main</artifactId>
-  <version>0.0.1-SNAPSHOT</version>
+  <version>0.1.0-SNAPSHOT</version>
 
 
 
@@ -21,8 +21,8 @@ Copyright (c) 2015-2016 Jorge Nunes, All Rights Reserved.
 
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 
-    <!-- By default unit tests are not run. -->
-    <skipTests>true</skipTests>
+    <!-- By default unit tests are run. -->
+    <skipTests>false</skipTests>
 
     <!-- Versions of dependencies and Maven plugins. -->
     <asciidoctorj.version>1.5.4</asciidoctorj.version>
@@ -273,7 +273,7 @@ Copyright (c) 2015-2016 Jorge Nunes, All Rights Reserved.
           <use>true</use>
           <quiet>true</quiet>
           <failOnError>false</failOnError>
-          <aggregate>true</aggregate>
+          <aggregate>false</aggregate>
         </configuration>
         <reportSets>
           <reportSet>
