+++ modules/release/pom.xml
@@ -15,7 +15,7 @@ Maven POM for the "release" sub-project of the Yawg project.
   <parent>
     <groupId>com.varmateo.yawg</groupId>
     <artifactId>yawg-main</artifactId>
-    <version>0.0.1-SNAPSHOT</version>
+    <version>0.1.0-SNAPSHOT</version>
     <relativePath>../</relativePath>
   </parent>
