+++ modules/testutils/pom.xml
@@ -15,7 +15,7 @@ Copyright (c) 2016 Jorge Nunes, All Rights Reserved.
   <parent>
     <groupId>com.varmateo.yawg</groupId>
     <artifactId>yawg-main</artifactId>
-    <version>0.0.1-SNAPSHOT</version>
+    <version>0.1.0-SNAPSHOT</version>
     <relativePath>..</relativePath>
   </parent>
